+++ OneBlink.SDK/models/Form.cs
@@ -322,9 +322,5 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public FormFooter footer
-        {
-            get; set;
-        }
     }
 }